fix(settings): 拆分恢复默认共享定义
